chore(flake): use runtime 9.0 instead of sdk to fix dotnet run bug
There seems to be bug either in nixpkgs packaging or in dotnet where
dotnet run will try to find files that do not exist when pointed to a
project. It will exit with No such file or directory. The files being
searched for can be seen in strace output. They are some pdb files in
the nix store.