feat(slashcommands): Make registering and removing commands more robust
When individual command creations/updates fail, there is no
information about what command failed. Additionally if that happens,
other commands don't get registered.

This makes it so that the application continues in registering slash
commands even if one errors. Then it aggregates the errors with
possibility of logging of all of them.